% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_BayesMDL.R
\docType{package}
\name{BayesMDL}
\alias{BayesMDL}
\alias{BayesMDL-package}
\title{Bayesian MDL for Multiple Changepoint Detection}
\description{
This package implements the Bayesian MDL Multiple Changepoint Detection
methods, which can be applied to univariate and bivariate time series data
and can accommodate metadata if exist.
}
