% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcp.R
\name{bcp}
\alias{bcp}
\title{Fit the changepoint detection method to multiple variables}
\usage{
bcp(X, pdf_file = "bcp.pdf", select = "MAP", iter = 10000, thin = max(1,
  iter/1000), weights = NULL, p = 2, time_unit = "month",
  seasonal_means = "harmonic", k = 3, scale_trend_design = 0.05,
  fit = "marlik", penalty = "bmdl", nu = 5, a = 1, b_eta = length(x),
  b_xi = length(x), max_changes = NULL, max_outliers = NULL,
  detect_outliers = TRUE, width = 8, height = 4, mar = c(5, 5, 4, 1),
  mfrow = c(1, 1), start_eta = NULL, track_time = TRUE)
}
\arguments{
\item{X}{Variables to be monitored, a \code{data.frame} of numerics, with
each row being a variable, in the format of a time series of length
\code{n}. If such information is in a CSV file, \code{X} can also be the
name of this file, with extension name, and a path if needed.}

\item{pdf_file}{Name of the output pdf file with extension name, a character;
can include path.}

\item{select}{The selected model to plot, \code{'MAP'} for maximum a
posterior, or \code{'BMA'} for median probability model under Bayesian
model averaging.}

\item{iter}{Total number of MCMC iterations.}

\item{thin}{Thinning; save one MCMC iteration for every \code{thin} number of
iterations.}

\item{weights}{A numeric vector of observation weights, defined the same as
the \code{weights} argument in the function \code{lm}.}

\item{p}{The order of the AR process.}

\item{time_unit}{Default is \code{'month'} (\code{'week'} may be added in the
future).}

\item{seasonal_means}{The seasonal means variables in the linear model.
Either \code{'fixed_effects'} such that each season has a coefficient
(with no intercept), or \code{'harmonic'} for harmonic regression (with
an intercept).}

\item{k}{The highest degree of harmonic regression. It is only used if
the argument \code{seasonal_means == 'harmonic'}.}

\item{scale_trend_design}{The factor multiplied to the design matrix of trend.
Default is 1/50.}

\item{fit}{For likelihood calculation, \code{'marlik'} for marginal
likelihood, or \code{'lik'} for likelihood. Note that the \code{'lik'}
option already includes the two-part MDL of \code{mu}.}

\item{penalty}{For penalty function calculation, \code{'bmdl'} for
Beta-Binomial prior, or \code{'mdl'} for MDL.}

\item{nu}{Prior variance scale of \code{mu}; only used if
\code{fit == 'marlik'}.}

\item{a}{The first parameter in the Beta-Binomial prior of
\code{eta} and \code{xi}; only used if \code{penalty == 'bmdl'}.}

\item{b_eta}{The second parameter in the Beta-Binomial prior of
\code{eta} and \code{xi}, respectively; only used if 
\code{penalty == 'bmdl'}.}

\item{b_xi}{The second parameter in the Beta-Binomial prior of
\code{eta} and \code{xi}, respectively; only used if 
\code{penalty == 'bmdl'}.}

\item{max_changes}{The maximum number of changepoints, or \code{NULL} if not
specified.}

\item{max_outliers}{The maximum number of outliers, or \code{NULL} if not
specified.}

\item{detect_outliers}{Logical, whether to detect outliers.}

\item{width, height}{Width and height of the output pdf (per page).}

\item{mar}{Margin of the output pdf (per page).}

\item{mfrow}{Number of rows and columns of figures on per pdf page.}

\item{start_eta}{A vector of 0/1 indicators for the initial model, or
\code{NULL} to randomly sample an initial model.}

\item{track_time}{Logical, whether to show runtime on screen.}
}
\value{
\item{eta_all}{Estimated changepoints for all variables; a matrix of 0/1
    indicators, where each row is for one variable, in the same order as
    the input data \code{X}.}
  \item{seasonal_cycles_all}{Estimated seasonal means for all variables,
    where each row is for one variable.}
  \item{mean_and_trend_all}{Estimated linear trends for all variables,
    where each row is for one variable.}
}
\description{
Apply Bayesian Minimum Description Length to detect multiple changepoints,
individually for a group of time series data (one for each variable to be
monitored). Returns the optimal changepoint models, and graphic outputs as a
pdf file.
}
\keyword{internal}
