% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdl.eta.R
\name{bmdl.eta}
\alias{bmdl.eta}
\title{Calculate BMDL for a Given Changepoint Model}
\usage{
bmdl.eta(X, month, eta, meta, p, fit, penalty, nu, a, b1, b2, period)
}
\arguments{
\item{X}{A vector of length \code{N}, observed time series data.}

\item{month}{A vector of length \code{N}, month indicators. Takes value in
{1, 2, ..., 12} for monthly data, or {1} for annual data.}

\item{eta}{A 0-1 indicator vector of length \code{N}, changepoint model. The
first \code{p} elemenet are always 0.}

\item{meta}{A 0-1 indicator vector of length \eqn{N}, metadata indicator.
The first \code{p} elemenet are always 0.}

\item{p}{The order of the AR process.}

\item{fit}{For likelihood calculation, \code{'marlik'} for marginal
likelihood, or \code{'lik'} for likelihood. Note that the \code{'lik'}
option already includes the two-part MDL of \code{mu}.}

\item{penalty}{For penalty function calculation, \code{'bmdl'} for
Beta-Binomial prior, or \code{'mdl'} for MDL.}

\item{nu}{Prior variance scale of \code{mu}; only used if
\code{fit == 'marlik'}.}

\item{a}{The first and second parameters in the Beta-Binomial prior; only
used if \code{penalty == 'bmdl'}.}

\item{b1}{The \eqn{b} parameter in the Beta-Binomial prior of \eqn{\eta}, for
un-documented times (i.e., times not in metadata). Default value is 239 for
monthly data, and 19 for annual data.}

\item{b2}{The \eqn{b} parameter in the Beta-Binomial prior of \eqn{\eta}, for
documented times (i.e., times in metadata). Default value is 47 for monthly
data, and 3 for annual data.}

\item{period}{Period of the seasonal cycle, 12 for monthly data, or 1 for
annual data.}
}
\value{
\item{mdl}{The Bayesian MDL score for the changepoint model \code{eta}.}
\item{phi}{A vector of length \code{p}, the Yule-Walker estimator of the
  autocorrelation parameter \code{phi}.}
\item{sigmasq}{The EB estimator of the white noise variance \code{sigmasq}.}
\item{mu}{A vector of length \code{sum(eta)}; the conditional posterior mean
  of regime-wise means \code{mu}.}
\item{s}{A vector of length \code{period}, the EB estimator of the monthly
  means \code{s}. }
}
\description{
For mean shifts detection in a univariate time series. These functions can
handle metadata.
}
\details{
\itemize{
\item To compute BMDL, use \code{fit = 'marlik'} and \code{penalty = 'bmdl'}.
\item To compute automatic MDL, use \code{fit = 'lik'} and
  \code{penalty = 'mdl'}.
\item To compute BIC, use \code{penalty = 'BIC'} (the \code{fit} argument
  does not matter in this case).
\item Note, the marginal likelihood \code{fit = 'marlik'} can also be
  combined with \code{penalty = 'uniform'}, which assigns uniform prior on
  the model space.
}
}
