% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdl.R
\name{bmdl.mean.shift}
\alias{bmdl.mean.shift}
\title{BMDL for Changepoint Detection}
\usage{
bmdl.mean.shift(X, month = NULL, meta = NULL, iter = 10000,
  thin = max(1, iter/1000), type = "monthly", p = 3, fit = "marlik",
  penalty = "bmdl", nu = 5, a = 1, b1 = 19 * (type == "annual") + 239 *
  (type == "monthly"), b2 = (b1 - 4)/5, start.eta = NULL,
  track.time = TRUE, show.summary = 10, show.month = FALSE,
  start.year = 1)
}
\arguments{
\item{X}{A vector of length \code{N}, observed time series data.}

\item{month}{A vector of length \code{N}, month indicators. Takes value in
{1, 2, ..., 12} for monthly data, or {1} for annual data.}

\item{meta}{A 0-1 indicator vector of length \eqn{N}, metadata indicator.
The first \code{p} elemenet are always 0.}

\item{iter}{Total number of MCMC iterations.}

\item{thin}{Thinning; save one MCMC iteration for every \code{thin} number of
iterations.}

\item{type}{Whether data are observed 'annual' or 'monthly'. Default is
'monthly'.}

\item{p}{The order of the AR process.}

\item{fit}{For likelihood calculation, \code{'marlik'} for marginal
likelihood, or \code{'lik'} for likelihood. Note that the \code{'lik'}
option already includes the two-part MDL of \code{mu}.}

\item{penalty}{For penalty function calculation, \code{'bmdl'} for
Beta-Binomial prior, or \code{'mdl'} for MDL.}

\item{nu}{Prior variance scale of \code{mu}; only used if
\code{fit == 'marlik'}.}

\item{a}{The first and second parameters in the Beta-Binomial prior; only
used if \code{penalty == 'bmdl'}.}

\item{b1}{The \eqn{b} parameter in the Beta-Binomial prior of \eqn{\eta}, for
un-documented times (i.e., times not in metadata). Default value is 239 for
monthly data, and 19 for annual data.}

\item{b2}{The \eqn{b} parameter in the Beta-Binomial prior of \eqn{\eta}, for
documented times (i.e., times in metadata). Default value is 47 for monthly
data, and 3 for annual data.}

\item{start.eta}{A 0-1 indicator vector of length \code{N}, initial
changepoint model for MCMC. The first \code{p} elemenet are always 0.
Default value is \code{NULL}.}

\item{track.time}{Logical; whether to report processing time (in seconds).}

\item{show.summary}{Integer; the number of top models to print. Default value
is 10.}

\item{show.month}{Logical; if show month in the summary. Default is
\code{FALSE}.}

\item{start.year}{Integer; the year time 1 is in. Default value is 1.}
}
\value{
\item{mcmc}{A (\code{iter / thin} + 1) by (\code{N} + \code{p} + 2) matrix;
  each row is a saved MCMC output from an iteration, consisting \code{eta},
  bmdl score, \code{phi}, and \code{sigmasq}.}
\item{map200}{A 200 by (\code{N} + \code{p} + 2) matrix; the best 200
  changepoint models visited. Each row is a model, consisting \code{eta},
  bmdl score, \code{phi}, and \code{sigmasq}.}
\item{input.parameters}{A list of all the following input parameters:
  \code{X}, \code{month}, \code{meta}, \code{iter}, \code{thin}, \code{type},
  \code{p}, \code{nu}, \code{a}, \code{b1}, \code{b2}, \code{period}, and
  \code{start.year}.}
}
\description{
For univiariate time series. It accomodates metadata (if any), and can handel
monthly data and annual data.
}
\details{
\itemize{
\item To compute BMDL, use \code{fit = 'marlik'} and \code{penalty = 'bmdl'}.
\item To compute automatic MDL, use \code{fit = 'lik'} and
  \code{penalty = 'mdl'}.
\item To compute BIC, use \code{penalty = 'BIC'} (the \code{fit} argument
  does not matter in this case).
\item Note, the marginal likelihood \code{fit = 'marlik'} can also be
  combined with \code{penalty = 'uniform'}, which assigns uniform prior on
  the model space.
}

Please see the references for more details.
}
\examples{
data(tuscaloosa);
X = tuscaloosa[, 3];

## For illustration purpose, here iter is small.
## To get meaningfull inference, use a larger value, e.g., iter = 1e5.
results = bmdl.mean.shift(X, month = tuscaloosa[, 2], meta = tuscaloosa[, 7],
                          iter = 1e2);

}
\references{
Yingbo Li, Robert Lund, and Anuradha Hewaarachchi,
"Multiple Changepoint Detection with Partial Information on Changepoint Times".
Working paper.
}
