% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdl.MH.R
\name{bmdl.MH}
\alias{bmdl.MH}
\alias{bmdl.MH.flip}
\alias{bmdl.MH.swap}
\title{Metropolis-Hastings Algorithms for Stoachstic Model Search}
\usage{
bmdl.MH.flip(X, month, current, meta, p, fit, penalty, nu, a, b1, b2, period)

bmdl.MH.swap(X, month, current, meta, p, fit, penalty, nu, a, b1, b2, period)
}
\arguments{
\item{X}{A vector of length \code{N}, observed time series data.}

\item{month}{A vector of length \code{N}, month indicators. Takes value in
{1, 2, ..., 12} for monthly data, or {1} for annual data.}

\item{current}{A list of current changepoint model. Specifically,
\code{list(eta, inference, change.eta)}, where \code{inference} is the
output of the function \code{\link{bmdl.eta}}.}

\item{meta}{A 0-1 indicator vector of length \eqn{N}, metadata indicator.
The first \code{p} elemenet are always 0.}

\item{p}{The order of the AR process.}

\item{fit}{For likelihood calculation, \code{'marlik'} for marginal
likelihood, or \code{'lik'} for likelihood. Note that the \code{'lik'}
option already includes the two-part MDL of \code{mu}.}

\item{penalty}{For penalty function calculation, \code{'bmdl'} for
Beta-Binomial prior, or \code{'mdl'} for MDL.}

\item{nu}{Prior variance scale of \code{mu}; only used if
\code{fit == 'marlik'}.}

\item{a}{The first and second parameters in the Beta-Binomial prior; only
used if \code{penalty == 'bmdl'}.}

\item{b1}{The \eqn{b} parameter in the Beta-Binomial prior of \eqn{\eta}, for
un-documented times (i.e., times not in metadata). Default value is 239 for
monthly data, and 19 for annual data.}

\item{b2}{The \eqn{b} parameter in the Beta-Binomial prior of \eqn{\eta}, for
documented times (i.e., times in metadata). Default value is 47 for monthly
data, and 3 for annual data.}

\item{period}{Period of the seasonal cycle, 12 for monthly data, or 1 for
annual data.}
}
\value{
A list of current changepoint model. Specifically,
  \code{list(eta, inference, change.eta)}, where \code{inference} is the
  output of the function \code{\link{bmdl.eta}}.
}
\description{
Two proposals: the function \code{bmdl.MH.flip} proposes to flip a random
  dimension, and the function \code{bmdl.MH.swap} propose to swap between a
  randomly chosen zero and a randomly chosen one.
}
