% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_updates.R
\name{eta_MH_birth}
\alias{eta_MH_birth}
\title{One MCMC iteration: propose to flip one random non-outlier component of eta
from 0 to 1}
\usage{
eta_MH_birth(x, A, current, p, fit, penalty, nu, kappa, a, b_eta, b_xi,
  scale_trend_design, weights, max_changes, eta_mcmc_prob)
}
\arguments{
\item{x}{The time series data, a numeric vector of length \code{n}.}

\item{A}{The design matrix for the nuisance coefficients in the linear model.
It is usually the matrix of seasonal indicators, or the design matrix
for harmonic regression with a column of all 1 for intercept.}

\item{current}{A list object representing a changepoint model. It contains
the following components: \code{eta}, \code{xi}, \code{inference} (output of
the \code{\link{fit_eta}} function), \code{change_eta}, and \code{change_xi}.}

\item{p}{The order of the AR process.}

\item{fit}{For likelihood calculation, \code{'marlik'} for marginal
likelihood, or \code{'lik'} for likelihood. Note that the \code{'lik'}
option already includes the two-part MDL of \code{mu}.}

\item{penalty}{For penalty function calculation, \code{'bmdl'} for
Beta-Binomial prior, or \code{'mdl'} for MDL.}

\item{nu}{Prior variance scale of \code{mu}; only used if
\code{fit == 'marlik'}.}

\item{kappa}{Prior variance scale of outliers.}

\item{a}{The first parameter in the Beta-Binomial prior of
\code{eta} and \code{xi}; only used if \code{penalty == 'bmdl'}.}

\item{b_eta}{The second parameter in the Beta-Binomial prior of
\code{eta} and \code{xi}, respectively; only used if 
\code{penalty == 'bmdl'}.}

\item{b_xi}{The second parameter in the Beta-Binomial prior of
\code{eta} and \code{xi}, respectively; only used if 
\code{penalty == 'bmdl'}.}

\item{scale_trend_design}{The factor multiplied to the design matrix of trend.
Default is 1/50.}

\item{weights}{A numeric vector of observation weights, defined the same as
the \code{weights} argument in the function \code{lm}.}

\item{max_changes}{The maximum number of changepoints, or \code{NULL} if not
specified.}

\item{eta_mcmc_prob}{A vector of three probablities, for the birth, death,
and swap proposals for the changepoint model \code{eta}. The three 
probabilities should add up to one.}
}
\value{
A list object representing the (maybe) updated changepoint model.
  \item{eta}{The (maybe) updated changepoint model, in the format of
    a vector of 0/1 indicators.}
  \item{xi}{The outliers, in the format of a vector of 0/1 indicators.}
  \item{inference}{Output of the \code{\link{fit_eta}} function on the output
    model \code{eta}, \code{xi}.}
  \item{change_eta}{Logical, if this \code{eta} is new, i.e.,
    the proposed model is accepted.}
  \item{change_xi}{Logical, if this \code{xi} is new, i.e.,
    the proposed model is accepted.}
}
\description{
One MCMC iteration: propose to flip one random non-outlier component of eta
from 0 to 1
}
\keyword{internal}
