% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_eta.R
\name{fit_eta}
\alias{fit_eta}
\title{Fit an individual changepoint model \code{eta} and outlier model \code{xi} 
pair}
\usage{
fit_eta(x, A, eta, xi, p = 2, fit = "marlik", penalty = "bmdl", nu = 5,
  kappa = 3, a = 1, b_eta = length(x), b_xi = length(x),
  scale_trend_design = 0.05, weights = NULL)
}
\arguments{
\item{x}{The time series data, a numeric vector of length \code{n}.}

\item{A}{The design matrix for the nuisance coefficients in the linear model.
It is usually the matrix of seasonal indicators, or the design matrix
for harmonic regression with a column of all 1 for intercept.}

\item{eta}{A multiple changepoint configuration, i.e., model. It is a 0/1
indicator vector of length \code{n}, with the first \code{max(p, 1)} 
elemenets always being 0. The maxinum number of 1's in \code{eta} is 
\code{ceiling((n-p-2*k-2)/2)-1}.}

\item{xi}{Outliers indicator, a 0/1 vector of length \code{n}.}

\item{p}{The order of the AR process.}

\item{fit}{For likelihood calculation, \code{'marlik'} for marginal
likelihood, or \code{'lik'} for likelihood. Note that the \code{'lik'}
option already includes the two-part MDL of \code{mu}.}

\item{penalty}{For penalty function calculation, \code{'bmdl'} for
Beta-Binomial prior, or \code{'mdl'} for MDL.}

\item{nu}{Prior variance scale of \code{mu}; only used if
\code{fit == 'marlik'}.}

\item{kappa}{Prior variance scale of outliers.}

\item{a}{The first parameter in the Beta-Binomial prior of
\code{eta} and \code{xi}; only used if \code{penalty == 'bmdl'}.}

\item{b_eta, b_xi}{The second parameter in the Beta-Binomial prior of
\code{eta} and \code{xi}, respectively; only used if 
\code{penalty == 'bmdl'}.}

\item{scale_trend_design}{The factor multiplied to the design matrix of trend.
Default is 1/50.}

\item{weights}{A numeric vector of observation weights, defined the same as
the \code{weights} argument in the function \code{lm}.}
}
\value{
\item{bmdl}{BMDL or MDL.}
\item{s}{Estimates of nuisance coefficients in the linear model. It usually
  contains the seasonal means or coefficients of harmonic regression
  (with intercept).}
\item{mu}{Estimates of regime-wise coefficients in the linear model. It
  usually contains the regime means and regime trends.}
\item{sigmasq}{Estimate of \eqn{\sigma^2} in the AR(p) process.}
\item{phi}{Yule-Walker estimates of the AR(p) coefficients.}
}
\description{
Compute the objective function of the model, i.e., log of posterior proablity
for BMDL, or log MDL, and estimate model parameters.
}
\keyword{internal}
