% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{harmonic_lm}
\alias{harmonic_lm}
\title{Harmonic regression}
\usage{
harmonic_lm(x, k = 3, time_ind = NULL, intercept = TRUE,
  return_design = TRUE)
}
\arguments{
\item{x}{Input time series of length \code{n}.}

\item{k}{The highest order of the factor inside sin/cos.}

\item{time_ind}{A index vector of times, or \code{NULL}, in which case
\code{1:n} will be used.}

\item{intercept}{Logical, whether to include an intercept.}

\item{return_design}{Logical, whether to return the design matrix.}
}
\value{
A linear regession \code{lm} object.
}
\description{
Harmonic regression
}
\keyword{internal}
