% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_bar_heat}
\alias{plot_bar_heat}
\title{Plot the heatmap and barplot of changepoints among multiple variables}
\usage{
plot_bar_heat(best_all, dates, type = "both")
}
\arguments{
\item{best_all}{A matrix of 0/1 indicators on the selected changepoint models.
Each row is the \code{eta} of a variable.}

\item{dates}{The dates records are observed, a \code{POSIXlt} vector. It should
have the \code{mon} component for months, and the \code{year} component
for years. See also \code{\link{strptime}}.}

\item{type}{Type of plot: \code{'heatmap'}, \code{'barplot'}, or \code{'both'}.}
}
\description{
Plot the heatmap and barplot of changepoints among multiple variables
}
\keyword{internal}
