% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_best}
\alias{plot_best}
\title{Plot eta of the MAP or BMA model of a MCMC fit}
\usage{
plot_best(results, select = "MAP", varname = NULL, plot_res = FALSE, ...)
}
\arguments{
\item{results}{Output object of the function \code{\link{bmdl}}.}

\item{select}{The selected model to plot, \code{'MAP'} for maximum a
posterior, or \code{'BMA'} for median probability model under Bayesian
model averaging.}

\item{varname}{The name of the variable \code{x}, a character.}

\item{plot_res}{Logical, whether residual plots will be added after the main
plots. If \code{TRUE}, then the following four plots will be added:
scatterplot and acf plot of the linear model residuals, and scatterplot
and acf plot fo the white noise estimates.}

\item{...}{Other arguments for the \code{\link{plot}} function.}
}
\description{
Plot eta of the MAP or BMA model of a MCMC fit
}
\keyword{internal}
