% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_res}
\alias{plot_res}
\title{Plot the residual}
\usage{
plot_res(current, x, dates, A, scale_trend_design, p, ...)
}
\arguments{
\item{current}{A list object representing a changepoint model. It contains
the following components: \code{eta}, \code{xi}, \code{inference} (output of
the \code{\link{fit_eta}} function), \code{change_eta}, and \code{change_xi}.}

\item{x}{The time series data, a numeric vector of length \code{n}.}

\item{dates}{The dates records are observed, a \code{POSIXlt} vector. It should
have the \code{mon} component for months, and the \code{year} component
for years. See also \code{\link{strptime}}.}

\item{A}{The design matrix for the nuisance coefficients in the linear model.
It is usually the matrix of seasonal indicators, or the design matrix
for harmonic regression with a column of all 1 for intercept.}

\item{scale_trend_design}{The factor multiplied to the design matrix of trend.
Default is 1/50.}

\item{p}{The order of the AR process.}

\item{...}{Other arguments for the \code{\link{plot}} function.}
}
\description{
Plot the residual
}
\keyword{internal}
