% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tuscaloosa}
\alias{tuscaloosa}
\title{Monthly Average Records of Daily Maximum and Minimum Temperatures in
Tuscaloosa, AL}
\format{A data frame with 1368 observations on the following 7 variables.
\describe{
  \item{year}{A numeric vector of year.}
  \item{month}{A numeric vector of month.}
  \item{target.max}{A numeric vector of the maximum temperatures of
    Tuscaloosa.}
  \item{target.min}{A numeric vector of the minimum temperatures of
    Tuscaloosa.}
  \item{ref.mean.max}{A numeric vector of the maximum temperatures of the
    reference series.}
  \item{ref.mean.min}{A numeric vector of the minimum temperatures of the
    reference series.}
  \item{meta}{a 0-1 vector of the Tuscaloosa station metadata.}
}}
\usage{
tuscaloosa
}
\description{
Reference series: average records of three nearby stations.
}
\keyword{datasets}
