% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txtn.eta.R
\name{txtn.eta}
\alias{txtn.eta}
\title{Calculate BMDL for a Given Changepoint Model, in Tmax / Tmin Joint Detection}
\usage{
txtn.eta(X, month, eta, meta, p, nu, alpha1, alpha2, period)
}
\arguments{
\item{X}{A \code{N} by 2 matrix, observed time series data. Columns are Tmax
and Tmin, respectively.}

\item{month}{A vector of length \code{N}, month indicators. Takes value in
{1, 2, ..., 12} for monthly data, or {1} for annual data.}

\item{eta}{A \code{N} by 3 matrix, changepoint model. The first two columns
are 0-1 indicators of changepoints for tmax and tmin, respectively, and the
third column is the category 1-4 each time is in. The first \code{p} times
are not allowed to be changepoints.}

\item{meta}{A 0-1 indicator vector of length \code{N}, metadata indicator.
The first \code{p} elemenet are always 0.}

\item{p}{The order of the VAR process.}

\item{nu}{Prior variance scale of \code{mu}; only used if
\code{fit == 'marlik'}.}

\item{alpha1}{The parameter in the Multinomial-Dirichlet prior of \code{eta},
for un-documented times (i.e., times not in metadata). Default value is
\code{c(3/7, 2/7, 2/7, 239)} for monthly data, and
\code{c(3/7, 2/7, 2/7, 19)} for annual data.}

\item{alpha2}{The parameter in the Multinomial-Dirichlet prior of \code{eta},
for documented times (i.e., times in metadata). Default value is
\code{c(3/7, 2/7, 2/7, 47)} for monthly data, and \code{c(3/7, 2/7, 2/7, 3)}
for annual data.}

\item{period}{Period of the seasonal cycle, 12 for monthly data and 1 for
annual data.}
}
\value{
\item{mdl}{The Bayesian MDL score for the changepoint model \code{eta}.}
\item{Phi}{A 2\code{p} by 2 matrix, the Yule-Walker estimator of the
  autocorrelation parameters \code{Phi_1}, ..., \code{Phi_p}, in a block
  matrix representation.}
\item{Sigma}{A 2 by 2 matrix, the Yule-Walker estimator of the white noise
  covariance \code{Sigma}.}
\item{mu}{A vector of length \code{sum(eta[ , 1]) + sum(eta[ , 2])}; the
  conditional posterior mean of regime-wise means \code{(mu_1, mu_2)}.}
\item{s}{A vector of length \code{period}, the EB estimator of the monthly
means \eqn{s}. }
}
\description{
For mean shifts detection in a bivariate time series. These functions can
handle metadata.
}
