% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txtn.MH.R
\name{txtn.MH}
\alias{txtn.MH}
\alias{txtn.MH.flip}
\alias{txtn.MH.swap}
\title{Metropolis-Hastings Algorithms for Stoachstic Model Search, in Tmax / Tmin
Joint Detection}
\usage{
txtn.MH.flip(X, month, current, meta, p, nu, alpha1, alpha2, period, q)

txtn.MH.swap(X, month, current, meta, p, nu, alpha1, alpha2, period)
}
\arguments{
\item{X}{A \code{N} by 2 matrix, observed time series data. Columns are Tmax
and Tmin, respectively.}

\item{month}{A vector of length \code{N}, month indicators. Takes value in
{1, 2, ..., 12} for monthly data, or {1} for annual data.}

\item{current}{A list of current changepoint model. Specifically,
\code{list(eta, count.eta, inference, change.eta)}, where \code{inference}
is the output of the function \code{\link{txtn.eta}}.}

\item{meta}{A 0-1 indicator vector of length \code{N}, metadata indicator.
The first \code{p} elemenet are always 0.}

\item{p}{The order of the VAR process.}

\item{nu}{Prior variance scale of \code{mu}; only used if
\code{fit == 'marlik'}.}

\item{alpha1}{The parameter in the Multinomial-Dirichlet prior of \code{eta},
for un-documented times (i.e., times not in metadata). Default value is
\code{c(3/7, 2/7, 2/7, 239)} for monthly data, and
\code{c(3/7, 2/7, 2/7, 19)} for annual data.}

\item{alpha2}{The parameter in the Multinomial-Dirichlet prior of \code{eta},
for documented times (i.e., times in metadata). Default value is
\code{c(3/7, 2/7, 2/7, 47)} for monthly data, and \code{c(3/7, 2/7, 2/7, 3)}
for annual data.}

\item{period}{Period of the seasonal cycle, 12 for monthly data and 1 for
annual data.}

\item{q}{A 4 by 4 transition matrix of the proposal distribution, only used
in the flip proposal.}
}
\value{
A list of current changepoint model. Specifically,
  \code{list(eta, count.eta, inference, change.eta)}, where \code{inference}
  is the output of the function \code{\link{txtn.eta}}.
}
\description{
Two proposals: the function \code{txtn.MH.flip} propose to flip a random
dimension, and the function \code{txtn.MH.swap} propose to swap between a
category 4 (no changes in either series) and a catogory 1-3 (change in at
least one series).
}
