% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txtn.R
\name{txtn}
\alias{txtn}
\title{BMDL for Changepoint Detection, in Tmax / Tmin Joint Detection}
\usage{
txtn(X, month = NULL, meta = NULL, iter = 10000, thin = max(1,
  iter/1000), type = "monthly", p = 3, nu = 5, alpha1 = c(3/7, 2/7, 2/7,
  19) * (type == "annual") + c(3/7, 2/7, 2/7, 239) * (type == "monthly"),
  alpha2 = c(3/7, 2/7, 2/7, 3) * (type == "annual") + c(3/7, 2/7, 2/7, 47) *
  (type == "monthly"), q = matrix(c(0, 1/2, 1/2, 1/2, 1/4, 0, 0, 1/4, 1/4, 0,
  0, 1/4, 1/2, 1/2, 1/2, 0), ncol = 4, nrow = 4), start.eta = NULL,
  track.time = TRUE, show.summary = 10, show.month = FALSE,
  start.year = 1)
}
\arguments{
\item{X}{A \code{N} by 2 matrix, observed time series data. Columns are Tmax
and Tmin, respectively.}

\item{month}{A vector of length \code{N}, month indicators. Takes value in
{1, 2, ..., 12} for monthly data, or {1} for annual data.}

\item{meta}{A 0-1 indicator vector of length \code{N}, metadata indicator.
The first \code{p} elemenet are always 0.}

\item{iter}{Total number of MCMC iterations.}

\item{thin}{Thinning; save one MCMC iteration for every \code{thin} number of
iterations.}

\item{type}{Whether data are observed 'annual' or 'monthly'. Default is
'monthly'.}

\item{p}{The order of the VAR process.}

\item{nu}{Prior variance scale of \code{mu}; only used if
\code{fit == 'marlik'}.}

\item{alpha1}{The parameter in the Multinomial-Dirichlet prior of \code{eta},
for un-documented times (i.e., times not in metadata). Default value is
\code{c(3/7, 2/7, 2/7, 239)} for monthly data, and
\code{c(3/7, 2/7, 2/7, 19)} for annual data.}

\item{alpha2}{The parameter in the Multinomial-Dirichlet prior of \code{eta},
for documented times (i.e., times in metadata). Default value is
\code{c(3/7, 2/7, 2/7, 47)} for monthly data, and \code{c(3/7, 2/7, 2/7, 3)}
for annual data.}

\item{q}{A 4 by 4 transition matrix of the proposal distribution, only used
in the flip proposal.}

\item{start.eta}{A 0-1 indicator vector of length \code{2*N}, initial
changepoint model for MCMC. The entries correspond to the first \code{p}
times are always 0. Default value is \code{NULL}.}

\item{track.time}{Logical; whether to report processing time (in seconds).}

\item{show.summary}{Integer; the number of top models to print. Default value
is 10.}

\item{show.month}{Logical; if show month in the summary. Default is
\code{FALSE}.}

\item{start.year}{Integer; the year time 1 is in. Default value is 1.}
}
\value{
\item{mcmc}{A (\code{iter / thin} + 1) by (2 * \code{N} + 4 * \code{p} + 5)
  matrix; each row is a saved MCMC output from an iteration, consisting
  \code{eta}, bmdl score, \code{Phi}, and \code{Sigma}.}
\item{map200}{A 200 by (2 *\code{N} + 4 * \code{p} + 5) matrix; the best 200
  changepoint models visited. Each row is a model, consisting \code{eta},
  bmdl score, \code{Phi}, and \code{Sigma}.}
\item{input.parameters}{A list of all the following input parameters:
  \code{X}, \code{month}, \code{meta}, \code{iter}, \code{thin}, \code{type},
  \code{p}, \code{nu}, \code{alpha1}, \code{alpha2}, \code{q}, \code{period},
  and \code{start.year}.}
}
\description{
For biviariate time series. It accomodates metadata (if any), and can handel
monthly data and annual data.
}
\examples{
data(tuscaloosa);
X = as.matrix(tuscaloosa[, 3:4]);

## For illustration purpose, here iter is small.
## To get meaningfull inference, use a larger value, e.g., iter = 5e4.
results = txtn(X, month = tuscaloosa[, 2], meta = tuscaloosa[, 7], iter = 1e1);
}
