% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdl_slope_outliers.R
\name{bmdl}
\alias{bmdl}
\title{Bayesian MDL (or MDL) for multiple changepoint detection}
\usage{
bmdl(x, dates, iter = 10000, thin = max(1, iter/1000), weights = NULL,
  p = 2, time_unit = "month", seasonal_means = "harmonic", k = 3,
  scale_trend_design = 0.05, fit = "marlik", penalty = "bmdl", nu = 5,
  kappa = 3, a = 1, b = 1, start_eta = NULL, start_xi = NULL,
  track_time = TRUE)
}
\arguments{
\item{x}{The time series data, a numeric vector of length \code{n}.}

\item{dates}{The dates records are observed, a \code{POSIXlt} vector. It should
have the \code{mon} component for months, and the \code{year} component
for years. See also \code{\link{strptime}}.}

\item{iter}{Total number of MCMC iterations.}

\item{thin}{Thinning; save one MCMC iteration for every \code{thin} number of
iterations.}

\item{weights}{A numeric vector of observation weights, defined the same as
the \code{weights} argument in the function \code{lm}.}

\item{p}{The order of the AR process.}

\item{time_unit}{Default is \code{'month'} (\code{'week'} may be added in the
future).}

\item{seasonal_means}{The seasonal means variables in the linear model.
Either \code{'fixed_effects'} such that each season has a coefficient
(with no intercept), or \code{'harmonic'} for harmonic regression (with
an intercept).}

\item{k}{The highest degree of harmonic regression. It is only used if
the argument \code{seasonal_means == 'harmonic'}.}

\item{scale_trend_design}{The factor multiplied to the design matrix of trend.
Default is 1/50.}

\item{fit}{For likelihood calculation, \code{'marlik'} for marginal
likelihood, or \code{'lik'} for likelihood. Note that the \code{'lik'}
option already includes the two-part MDL of \code{mu}.}

\item{penalty}{For penalty function calculation, \code{'bmdl'} for
Beta-Binomial prior, or \code{'mdl'} for MDL.}

\item{nu}{Prior variance scale of \code{mu}; only used if
\code{fit == 'marlik'}.}

\item{kappa}{Prior variance scale of outliers.}

\item{a}{The first and second parameters in the Beta-Binomial prior; only
used if \code{penalty == 'bmdl'}.}

\item{b}{The first and second parameters in the Beta-Binomial prior; only
used if \code{penalty == 'bmdl'}.}

\item{start_eta}{A vector of 0/1 indicators for the initial model, or
\code{NULL} to randomly sample an initial model.}

\item{start_xi}{A vector of 0/1 indicators for the initial outliers, or
\code{NULL} to randomly sample an initial model.}

\item{track_time}{Logical, whether to show runtime on screen.}
}
\value{
\item{best}{The optimal model which minimizes BMDL or MDL; a list representing
  the model, which contains components: \code{eta}, \code{inference} (output
  of the \code{\link{fit_eta}} function), and \code{change_eta}.}
\item{eta_mcmc}{A matrix to save MCMC iterations of \code{eta}. Each row
  is about an iteration, in the format of a vector of length \code{n + p + 2},
  containing \code{eta}, \code{bmdl}, \code{phi}, \code{sigmasq}.}
\item{A}{The design matrix for the nuisance coefficients in the linear model.
  It is usually the matrix of seasonal indicators, if the argument
  \code{seasonal_means = 'fixed_effects'}, or the design matrix
  for harmonic regression with a column of all 1 for intercept, if the
  argument \code{seasonal_means = 'harmonic'}).}
\item{runtime}{Runtime, in second.}
\item{input_parameters}{A list of input parameters.}
}
\description{
Changes in both mean and linear trend, which permitting a global seasonal
mean and AR(p) errors.
}
\keyword{internal}
