% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{last_cp}
\alias{last_cp}
\title{Location of the last changepoint}
\usage{
last_cp(eta)
}
\arguments{
\item{eta}{A multiple changepoint configuration, i.e., model. It is a 0/1
indicator vector of length \code{n}, with the first \code{max(p, 1)} 
elemenets always being 0. The maxinum number of 1's in \code{eta} is 
\code{ceiling((n-p-2*k-2)/2)-1}.}
}
\value{
An iterger, the location of the last changepoint, or \code{NA} if
  there are no changepoints.
}
\description{
Location of the last changepoint
}
\keyword{internal}
