% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_eta}
\alias{plot_eta}
\title{Plot time series data, changepoints, and outliers}
\usage{
plot_eta(x, dates, varname = NULL, current = NULL, A = NULL,
  scale_trend_design = NULL, eta = NULL, xi = NULL,
  seasonal_cycles = NULL, mean_and_trend = NULL, plot_res = FALSE,
  yw_phi = NULL, ...)
}
\arguments{
\item{x}{The time series data, a numeric vector of length \code{n}.}

\item{dates}{The dates records are observed, a \code{POSIXlt} vector. It should
have the \code{mon} component for months, and the \code{year} component
for years. See also \code{\link{strptime}}.}

\item{varname}{The name of the variable \code{x}, a character.}

\item{current}{A list object representing a changepoint model. It contains
the following components: \code{eta}, \code{xi}, \code{inference} (output of
the \code{\link{fit_eta}} function), \code{change_eta}, and \code{change_xi}.}

\item{A}{The design matrix for the nuisance coefficients in the linear model.
It is usually the matrix of seasonal indicators, or the design matrix
for harmonic regression with a column of all 1 for intercept.}

\item{scale_trend_design}{The factor multiplied to the design matrix of trend.
Default is 1/50.}

\item{eta}{A multiple changepoint configuration, i.e., model. It is a 0/1
indicator vector of length \code{n}, with the first \code{max(p, 1)} 
elemenets always being 0. The maxinum number of 1's in \code{eta} is 
\code{ceiling((n-p-2*k-2)/2)-1}.}

\item{xi}{Outliers indicator, a 0/1 vector of length \code{n}.}

\item{seasonal_cycles}{Estimated seasonal means, a numeric vector of length
\code{n}.}

\item{mean_and_trend}{Estimated piecewise linear fits, a numeric vector of
length \code{n}.}

\item{plot_res}{Logical, whether residual plots will be added after the main
plots. If \code{TRUE}, then the following four plots will be added:
scatterplot and acf plot of the linear model residuals, and scatterplot
and acf plot fo the white noise estimates.}

\item{yw_phi}{Yule walker estimator of AR coefficients, a vector of length
\code{p}; must be non-NULL if \code{plot_res == TRUE} and
\code{is.null(current)}.}

\item{...}{Other arguments for the \code{\link{plot}} function.}
}
\description{
Create a plot, with time series data \code{x} shown in points, linear trends
under the selected changepoint model in blue lines, estimated changepoints
in verticle dashed blue lines, and fitted value (including seasonal means
and linear trends) in a purple line. Optional residual plots can also be
plotted.
}
\details{
Two sets of inputs can be used:
\enumerate{
  \item Either \code{current}, \code{A}, \code{scale_trend_design},
  \item Or \code{eta}, \code{xi}, \code{seasonal_cycles}, 
  \code{mean_and_trend}.
}
When used, every inputs in that set must but non-NULL, while all inputs in
the other set must be NULL.
}
\keyword{internal}
